Feature/persistent subscriptions (#8)
Refactored the messaging module. We now have `messages` which contains low-level types representing the bytes on the network, and the `conversations` module which represents sequences of messages. 

Conversations and messages work without any io, so they're easier to test. This work flushed out a whole bunch of bugs and error handling.