Bugfix/dispatcher should not block heartbeats (#58)
Fixes #49 

This commit introduces a new class PaceMaker whose job is to send and receive heartbeat messages. Taking this responsibility away from the MessageDispatcher stops us timing out heartbeats just because a message is taking time to process in the application.

We can still fail to receive heartbeats if we receive large messages from the server. There's almost certainly a ton of performance tuning we could do inside the MessageReader to help with this case.

Minor Tweaks:

    When paging through a stream with the IterStreamEvents conversation, we now request page N+1 before yielding events from page N; this means less time waiting for responses because we can receive the next page before we've finished processing the last one.
    Added an ugly conditional so that we don't waste time formatting a dump unless we're actually running in trace mode.