subscribe_to now takes an optional start_from
By default, start_from is -1, which is a const used elsewhere for
"beginning of stream". Any other value will cause PhotonPump to
create a catch-up subscription and read historical events from
the `start_from` event number.

Performed a quick test-firing, but we still need to implement
reconnection logic before this is ready for master.